--- src/Pages/Moods.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-function Moods() {
-  return (
-    <div className='moods' >
-      <h1>Moods</h1>
-    </div>
-  );
-}
-
-export default Moods;
